requirements.txt

streamlit==1.26.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.30.0
pandas==1.5.3